% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-create.R
\name{bdc_data_products}
\alias{bdc_data_products}
\title{Returns a tibble with a product ID and a randomly assigned price}
\usage{
bdc_data_products(
  no_products = 30,
  seed = 7878,
  price_low = 4,
  price_high = 10
)
}
\arguments{
\item{no_products}{Number of products to produce}

\item{seed}{Seed number to use for random data}

\item{price_low}{Lowest price to assign}

\item{price_high}{Highest price to assign}
}
\description{
Returns a tibble with a product ID and a randomly assigned price
}
