% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-init.R
\name{bdc_db_init}
\alias{bdc_db_init}
\title{Initializes schema, tables and views}
\usage{
bdc_db_init(con, schema_name = "retail", overwrite = FALSE)
}
\arguments{
\item{con}{PostgreSQL database connection}

\item{schema_name}{Name of the schema to use}

\item{overwrite}{If exists, it will delete the existing schema and replace it with the new content}
}
\description{
Initializes schema, tables and views
}
