% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-init.R
\name{bdc_db_sqlite}
\alias{bdc_db_sqlite}
\title{Creates an empty SQLite database}
\usage{
bdc_db_sqlite(folder = "database")
}
\arguments{
\item{folder}{Folder location to place the database file}
}
\description{
Creates an empty SQLite database
}
