% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bdc_list_rmds}
\alias{bdc_list_rmds}
\title{Returns a vector with the path to each workbook}
\usage{
bdc_list_rmds(book_path = "workbook")
}
\arguments{
\item{book_path}{The path to the workbook files}
}
\description{
Returns a vector with the path to each workbook
}
