% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bdc_renumber_file}
\alias{bdc_renumber_file}
\title{Copies a workbook into another folder and re-numbers the steps}
\usage{
bdc_renumber_file(src_path, target_folder)
}
\arguments{
\item{src_path}{The path to the workbooks file}

\item{target_folder}{The folder address to save the new file to}
}
\description{
Copies a workbook into another folder and re-numbers the steps
}
