% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bdc_renumber_folder}
\alias{bdc_renumber_folder}
\title{Copies the workbooks into another folder and re-numbers the steps}
\usage{
bdc_renumber_folder(book_path = "workbook", target_folder = "")
}
\arguments{
\item{book_path}{The path to the workbook files}

\item{target_folder}{The folder address to save the new files to}
}
\description{
Copies the workbooks into another folder and re-numbers the steps
}
