% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bdc_utils_exercises}
\alias{bdc_utils_exercises}
\title{Copies the workbooks into another folder and re-numbers the steps}
\usage{
bdc_utils_exercises(book_path = "workbook", exercise_folder = "")
}
\arguments{
\item{book_path}{The path to the workbook files}

\item{exercise_folder}{The folder address to save the new files to}
}
\description{
Copies the workbooks into another folder and re-numbers the steps
}
